Add select all/deselect all competencies in group to verify multiple results view
