+++ DigitalLearningSolutions.Web/Scripts/learningPortal/selfAssessment.ts
@@ -1,3 +1,5 @@
+import * as CheckboxGroup from '../checkboxGroup';
+
 function onSliderUpdate(inputElement: HTMLInputElement) {
   const selectedRatio = parseInt(inputElement.value, 10) / parseInt(inputElement.max, 10);
   // eslint-disable-next-line no-param-reassign
@@ -14,42 +16,4 @@ inputs.forEach((e) => {
   e.addEventListener('change', () => onSliderUpdate(e));
 });
 
-setUpSelectAndDeselectButtons();
-
-function setUpSelectAndDeselectButtons(): void {
-  const selectAllButtons = document.querySelectorAll('.select-all') as NodeListOf<HTMLAnchorElement>;
-  selectAllButtons.forEach((button) => {
-    button.addEventListener('click',
-      () => {
-        const group = button.getAttribute('data-group') as string;
-        selectAll(group);
-      });
-  });
-
-  const deselectAllButtons = document.querySelectorAll('.deselect-all') as NodeListOf<HTMLAnchorElement>;
-  deselectAllButtons.forEach((button) => {
-    button.addEventListener('click',
-      () => {
-        const group = button.getAttribute('data-group') as string;
-        deselectAll(group);
-      });
-  });
-}
-
-function selectAll(group: string): void {
-  const allCheckboxes = document.querySelectorAll('.select-all-checkbox') as NodeListOf<HTMLInputElement>;
-  allCheckboxes.forEach((checkbox) => {
-    if (checkbox.getAttribute('data-group') === group) {
-      if (!checkbox.checked) checkbox.click();
-    }
-  });
-}
-
-function deselectAll(group: string): void {
-  const allCheckboxes = document.querySelectorAll('.select-all-checkbox') as NodeListOf<HTMLInputElement>;
-  allCheckboxes.forEach((checkbox) => {
-    if (checkbox.getAttribute('data-group') === group) {
-      if (checkbox.checked) checkbox.click();
-    }
-  });
-}
+CheckboxGroup.default.setUpSelectAndDeselectButtons();
