+++ DigitalLearningSolutions.Web/Scripts/supervisor/verifyMultipleResults.ts
@@ -1,0 +1,3 @@
+import * as CheckboxGroup from '../checkboxGroup';
+
+CheckboxGroup.default.setUpSelectAndDeselectButtons();
