+++ DigitalLearningSolutions.Web/Styles/learningPortal/selfAssessment.scss
@@ -4,12 +4,6 @@
 @import "courses";
 @import "../shared/searchableElements/searchableElements";
 
-.select-all-button {
-  @extend .right-submit-button;
-  width: auto;
-  border-radius: 4px;
-  margin-right: nhsuk-spacing(2);
-}
 .competency-group-header {
   padding-top: 0;
   margin-bottom: 4px;
