+++ DigitalLearningSolutions.Web/Styles/supervisor/verifyMultipleResults.scss
@@ -1,0 +1,3 @@
+﻿.nhsuk-grid-column-full {
+  height: 40px;
+}
